Removed OFFSET and LEADING-DIMENSION.
Rationale: LAPACK procedures sometimes return the result in one of the
arguments, and handling offset and leading-dimension either leads to
copying, or an elaborate and error-prone scheme of keeping track of
leading dimensions and offsets.  Better to keep it simple and write
fast copying methods.

This resulted in other changes:

- removed compact-matrix
- row-adjustable-matrix is no longer a subclass of dense-matrix-like
- introduced dense-matrix-slots (mixin), defines xrank, xdims, xdim
- submatrix always copies, also becomes generic function
- abstracted out submatrix offset & index calculations
- removed leading-dimension from all with-matrix-input(s) calls
- added necessary unit tests