Changes to &info.
Added explicit variable for positive INFO arguments, which can now be
checked by the return form.

When CONDITION is NIL, positive INFO arguments are ignored.

This was necessary because some LAPACK return meaningful results even
for a positive INFO, which is then something like a warning (eg an LU decomposition of a singular matrix).